+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changelog #
 
+## Version 1.5.0
+
+Date: 2017-04-17
+
+- Add the ability to pass customized instance of `goog.history.Html5History` to be
+  used to manage history events.
+
 ## Version 1.4.0
 
 Date: 2017-01-12